[rename] hook들 폴더 경로 이동
