+++ apps/frontend/src/hooks/useLotteryCanvas.ts
