+++ apps/frontend/src/hooks/useLotteryModal.ts
